Move FT6X36 files and add its license
