+++ Boards/M5stackCore2/Source/ft6x36/README.md
@@ -1,0 +1,8 @@
+This project is an adaption of the code at https://github.com/martinberlin/FT6X36-IDF which is an adaptation of https://github.com/strange-v/FT6X36
+The original license is an MIT license and is included in this directory.
+
+Changes:
+- Remove Kconfig-based configuratio
+- Removed I2C init code
+- Allow for passing a different I2C port
+
